moved ef core tests to unit tests project
