+++ BookingsApi/BookingsApi.UnitTests/Domain/AnonymisationData/GetAnonymisationDataQueryHandlerTests.cs
@@ -5,13 +5,12 @@ using System.Threading.Tasks;
 using BookingsApi.DAL;
 using BookingsApi.DAL.Queries;
 using BookingsApi.Domain;
-using BookingsApi.Domain.Participants;
 using BookingsApi.Domain.RefData;
 using FluentAssertions;
 using Microsoft.EntityFrameworkCore;
 using NUnit.Framework;
 
-namespace BookingsApi.IntegrationTests.Database.Queries
+namespace BookingsApi.UnitTests.Domain.AnonymisationData
 {
     public class GetAnonymisationDataQueryHandlerTests
     {
